Remove writing to Zookeeper when kafka is selected as offset storage and dual.commit is false.

+++ src/main/java/com/pinterest/secor/uploader/Uploader.java
@@ -66,6 +66,7 @@ public class Uploader {
     protected String mTopicFilter;
 
     private boolean isOffsetsStorageKafka = false;
+    private boolean isOffsetsStorageZookeeper = false;
 
 
     /**
@@ -104,9 +105,13 @@ public class Uploader {
             isOffsetsStorageKafka = true;
         }
         mUploadLastSeenOffset = mConfig.getUploadLastSeenOffset();
+        if (mConfig.getOffsetsStorage().equals(SecorConstants.KAFKA_OFFSETS_STORAGE_ZK) || mConfig.getDualCommitEnabled().equals("true")) {
+            isOffsetsStorageZookeeper = true;
+        }
     }
 
     protected void uploadFiles(TopicPartition topicPartition) throws Exception {
+        LOG.debug("Uploading for: " + topicPartition);
         long committedOffsetCount = mOffsetTracker.getTrueCommittedOffsetCount(topicPartition);
         long lastSeenOffset = mOffsetTracker.getLastSeenOffset(topicPartition);
 
@@ -122,10 +127,9 @@ public class Uploader {
         mZookeeperConnector.lock(lockPath);
         try {
             // Check if the committed offset has changed.
-            long zookeeperCommittedOffsetCount = mZookeeperConnector.getCommittedOffsetCount(
-                    topicPartition);
-            if (zookeeperCommittedOffsetCount == committedOffsetCount) {
-                if (mUploadLastSeenOffset) {
+            long zookeeperCommittedOffsetCount = mZookeeperConnector.getCommittedOffsetCount(topicPartition);
+            if (zookeeperCommittedOffsetCount == committedOffsetCount || !isOffsetsStorageZookeeper) {
+                if (mUploadLastSeenOffset && isOffsetsStorageZookeeper) {
                     long zkLastSeenOffset = mZookeeperConnector.getLastSeenOffsetCount(topicPartition);
                     // If the in-memory lastSeenOffset is less than ZK's, this means there was a failed
                     // attempt uploading for this topic partition and we already have some files uploaded
@@ -171,11 +175,17 @@ public class Uploader {
                 if (mDeterministicUploadPolicyTracker != null) {
                     mDeterministicUploadPolicyTracker.reset(topicPartition);
                 }
-                mZookeeperConnector.setCommittedOffsetCount(topicPartition, lastSeenOffset + 1);
+
                 mOffsetTracker.setCommittedOffsetCount(topicPartition, lastSeenOffset + 1);
+
                 if (isOffsetsStorageKafka) {
                     mMessageReader.commit(topicPartition, lastSeenOffset + 1);
                 }
+
+                if (isOffsetsStorageZookeeper) {
+                    mZookeeperConnector.setCommittedOffsetCount(topicPartition, lastSeenOffset + 1);
+                }
+
                 mMetricCollector.increment("uploader.file_uploads.count", paths.size(), topicPartition.getTopic());
             } else {
                 LOG.warn("Zookeeper committed offset didn't match for topic {} partition {}: {} vs {}",
@@ -308,35 +318,38 @@ public class Uploader {
                            isRequiredToUploadAtTime(topicPartition);
         }
         if (shouldUpload) {
-            long newOffsetCount = mZookeeperConnector.getCommittedOffsetCount(topicPartition);
-            long oldOffsetCount = mOffsetTracker.setCommittedOffsetCount(topicPartition,
-                    newOffsetCount);
-            long lastSeenOffset = mOffsetTracker.getLastSeenOffset(topicPartition);
-            if (oldOffsetCount == newOffsetCount) {
-                LOG.debug("Uploading for: " + topicPartition);
+            if (!isOffsetsStorageZookeeper) {
                 uploadFiles(topicPartition);
-            } else if (newOffsetCount > lastSeenOffset) {  // && oldOffset < newOffset
-                LOG.debug("last seen offset {} is lower than committed offset count {}. Deleting files in topic {} partition {}",
-                        lastSeenOffset, newOffsetCount,topicPartition.getTopic(), topicPartition.getPartition());
-                mMetricCollector.increment("uploader.partition_deletes", topicPartition.getTopic());
-                // There was a rebalancing event and someone committed an offset beyond that of the
-                // current message.  We need to delete the local file.
-                mFileRegistry.deleteTopicPartition(topicPartition);
-                if (mDeterministicUploadPolicyTracker != null) {
-                    mDeterministicUploadPolicyTracker.reset(topicPartition);
+            } else {
+                long newOffsetCount = mZookeeperConnector.getCommittedOffsetCount(topicPartition);
+                long oldOffsetCount = mOffsetTracker.setCommittedOffsetCount(topicPartition,
+                        newOffsetCount);
+                long lastSeenOffset = mOffsetTracker.getLastSeenOffset(topicPartition);
+                if (oldOffsetCount == newOffsetCount) {
+                    uploadFiles(topicPartition);
+                } else if (newOffsetCount > lastSeenOffset) {  // && oldOffset < newOffset
+                    LOG.debug("last seen offset {} is lower than committed offset count {}. Deleting files in topic {} partition {}",
+                            lastSeenOffset, newOffsetCount, topicPartition.getTopic(), topicPartition.getPartition());
+                    mMetricCollector.increment("uploader.partition_deletes", topicPartition.getTopic());
+                    // There was a rebalancing event and someone committed an offset beyond that of the
+                    // current message.  We need to delete the local file.
+                    mFileRegistry.deleteTopicPartition(topicPartition);
+                    if (mDeterministicUploadPolicyTracker != null) {
+                        mDeterministicUploadPolicyTracker.reset(topicPartition);
+                    }
+                } else {  // oldOffsetCount < newOffsetCount <= lastSeenOffset
+                    LOG.debug("previous committed offset count {} is lower than committed offset {} is lower than or equal to last seen offset {}. " +
+                                    "Trimming files in topic {} partition {}",
+                            oldOffsetCount, newOffsetCount, lastSeenOffset, topicPartition.getTopic(), topicPartition.getPartition());
+                    mMetricCollector.increment("uploader.partition_trims", topicPartition.getTopic());
+                    // There was a rebalancing event and someone committed an offset lower than that
+                    // of the current message.  We need to trim local files.
+                    trimFiles(topicPartition, newOffsetCount);
+                    // We might still be at the right place to upload. (In fact, we always trim the first time
+                    // we hit the upload condition because oldOffsetCount starts at -1, but this is usually a no-op trim.)
+                    // Check again! This is especially important if this was an "upload in graceful shutdown".
+                    checkTopicPartition(topicPartition, forceUpload);
                 }
-            } else {  // oldOffsetCount < newOffsetCount <= lastSeenOffset
-                LOG.debug("previous committed offset count {} is lower than committed offset {} is lower than or equal to last seen offset {}. " +
-                                "Trimming files in topic {} partition {}",
-                        oldOffsetCount, newOffsetCount, lastSeenOffset, topicPartition.getTopic(), topicPartition.getPartition());
-                mMetricCollector.increment("uploader.partition_trims", topicPartition.getTopic());
-                // There was a rebalancing event and someone committed an offset lower than that
-                // of the current message.  We need to trim local files.
-                trimFiles(topicPartition, newOffsetCount);
-                // We might still be at the right place to upload. (In fact, we always trim the first time
-                // we hit the upload condition because oldOffsetCount starts at -1, but this is usually a no-op trim.)
-                // Check again! This is especially important if this was an "upload in graceful shutdown".
-                checkTopicPartition(topicPartition, forceUpload);
             }
         }
     }
